src/requirements.txt
Flask
Docker
Redis
Pillow